refactor(cervin-p5): rename cervinrc.ts → transitrixrc.ts; keep cervinrc.ts as re-export shim
- Creates src/transitrixrc.ts with canonical names: TRANSITRIXRC_SCHEMA,
  validateTransitrixrcDocument (both were private, no public-API break)
- Reduces src/cervinrc.ts to a @deprecated re-export stub pointing to
  transitrixrc.ts; preserves loadCervinrc alias through 1.x
- Updates src/compiler.ts to import from './transitrixrc.js'
- Renames tests/cervinrc.test.ts → tests/transitrixrc.test.ts and updates
  import path + describe label; re-export shim keeps the deprecated-alias
  test coverage intact from the old suite

Completes Cervin deprecation P5 (runtime rename of internal module).
